kiro2api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.2
python-dotenv>=1.0.0
pydantic>=2.10.0
json_repair>=0.48.0
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
sse-starlette>=1.6.5

# Camoufox 反检测浏览器自动化
camoufox[geoip]>=0.4.0
playwright>=1.40.0
pyotp>=2.9.0